Refactor odie-notice
