+++ packages/odie-client/src/components/message/messages-container.tsx
@@ -22,8 +22,8 @@ import {
 	hasCSATMessage,
 	hasSubmittedCSATRating,
 } from '../../utils';
-import { OdieNotices } from '../odie-notice';
-import useViewMostRecentOpenConversationNotice from '../odie-notice/use-view-most-recent-conversation-notice';
+import { Notices } from '../notices';
+import useViewMostRecentOpenConversationNotice from '../notices/use-view-most-recent-conversation-notice';
 import { JumpToRecent } from './jump-to-recent';
 import { ThinkingPlaceholder } from './thinking-placeholder';
 import ChatMessage from '.';
@@ -164,76 +164,74 @@ export const MessagesContainer = ( { currentUser }: ChatMessagesProps ) => {
 	const chatHasCSATMessage = hasCSATMessage( chat );
 	const displayCSAT = chatHasCSATMessage && ! hasSubmittedCSATRating( chat );
 	return (
-		<>
+		<div
+			className={ clx( 'chatbox-messages', {
+				'force-email-support': forceEmailSupport && chat.provider === 'zendesk',
+			} ) }
+			ref={ messagesContainerRef }
+		>
 			<div
-				className={ clx( 'chatbox-messages', {
-					'force-email-support': forceEmailSupport && chat.provider === 'zendesk',
-				} ) }
-				ref={ messagesContainerRef }
+				className="screen-reader-text"
+				aria-live="polite"
+				aria-atomic="false"
+				aria-relevant="additions"
 			>
-				<div
-					className="screen-reader-text"
-					aria-live="polite"
-					aria-atomic="false"
-					aria-relevant="additions"
-				>
-					{ chat.messages.map( ( message ) => (
-						<div key={ `${ message.internal_message_id }` }>
-							{ [ 'bot', 'business' ].includes( message.role ) && message.content }
-						</div>
-					) ) }
-				</div>
-				<ChatDate chat={ chat } />
-				{ ! chatMessagesLoaded ? (
-					<LoadingChatSpinner />
-				) : (
-					<>
-						{ ( chat.odieId || chat.provider === 'odie' ) && (
-							<ChatMessage
-								message={ getOdieInitialMessage( botNameSlug ) }
-								key={ 0 }
-								currentUser={ currentUser }
-								isNextMessageFromSameSender={ false }
-								displayChatWithSupportLabel={ false }
-							/>
-						) }
-						{ chat.messages.map( ( message, index ) => {
-							const nextMessage = chat.messages[ index + 1 ];
-							const displayChatWithSupportLabel =
-								! nextMessage?.context?.flags?.show_contact_support_msg &&
-								message.context?.flags?.show_contact_support_msg &&
-								! chatHasEnded &&
-								! message.context?.flags?.is_error_message;
-
-							const displayChatWithSupportEndedLabel =
-								! chatHasCSATMessage && ! nextMessage && chatHasEnded;
-
-							return (
-								<ChatMessage
-									message={ message }
-									key={ index }
-									currentUser={ currentUser }
-									isNextMessageFromSameSender={ isNextMessageFromSameSender(
-										message.role,
-										chat.messages[ index + 1 ]?.role
-									) }
-									displayChatWithSupportLabel={ displayChatWithSupportLabel }
-									displayChatWithSupportEndedLabel={ displayChatWithSupportEndedLabel }
-									displayCSAT={ displayCSAT }
-								/>
-							);
-						} ) }
-						<JumpToRecent containerReference={ messagesContainerRef } />
-
-						{ chat.provider === 'odie' && chat.status === 'sending' && (
-							<div className="odie-chatbox__action-message">
-								<ThinkingPlaceholder />
-							</div>
-						) }
-						<OdieNotices />
-					</>
-				) }
+				{ chat.messages.map( ( message ) => (
+					<div key={ `${ message.internal_message_id }` }>
+						{ [ 'bot', 'business' ].includes( message.role ) && message.content }
+					</div>
+				) ) }
 			</div>
-		</>
+			<ChatDate chat={ chat } />
+			{ ! chatMessagesLoaded ? (
+				<LoadingChatSpinner />
+			) : (
+				<>
+					{ ( chat.odieId || chat.provider === 'odie' ) && (
+						<ChatMessage
+							message={ getOdieInitialMessage( botNameSlug ) }
+							key={ 0 }
+							currentUser={ currentUser }
+							isNextMessageFromSameSender={ false }
+							displayChatWithSupportLabel={ false }
+						/>
+					) }
+					{ chat.messages.map( ( message, index ) => {
+						const nextMessage = chat.messages[ index + 1 ];
+						const displayChatWithSupportLabel =
+							! nextMessage?.context?.flags?.show_contact_support_msg &&
+							message.context?.flags?.show_contact_support_msg &&
+							! chatHasEnded &&
+							! message.context?.flags?.is_error_message;
+
+						const displayChatWithSupportEndedLabel =
+							! chatHasCSATMessage && ! nextMessage && chatHasEnded;
+
+						return (
+							<ChatMessage
+								message={ message }
+								key={ index }
+								currentUser={ currentUser }
+								isNextMessageFromSameSender={ isNextMessageFromSameSender(
+									message.role,
+									chat.messages[ index + 1 ]?.role
+								) }
+								displayChatWithSupportLabel={ displayChatWithSupportLabel }
+								displayChatWithSupportEndedLabel={ displayChatWithSupportEndedLabel }
+								displayCSAT={ displayCSAT }
+							/>
+						);
+					} ) }
+					<JumpToRecent containerReference={ messagesContainerRef } />
+
+					{ chat.provider === 'odie' && chat.status === 'sending' && (
+						<div className="odie-chatbox__action-message">
+							<ThinkingPlaceholder />
+						</div>
+					) }
+					<Notices />
+				</>
+			) }
+		</div>
 	);
 };
