+++ packages/odie-client/src/components/notices/index.tsx
@@ -2,12 +2,12 @@ import './style.scss';
 import { Icon, close } from '@wordpress/icons';
 import { useOdieAssistantContext } from '../../context';
 
-interface OdieNoticeProps {
+interface NoticeProps {
 	content?: string | React.ReactNode;
 	onClose?: () => void;
 }
 
-export const OdieNotice: React.FC< OdieNoticeProps > = ( { content, onClose } ) => {
+export const Notice: React.FC< NoticeProps > = ( { content, onClose } ) => {
 	return (
 		<div className="odie-notice">
 			{ content && <span>{ content }</span> }
@@ -18,7 +18,7 @@ export const OdieNotice: React.FC< OdieNoticeProps > = ( { content, onClose } ) 
 	);
 };
 
-export const OdieNotices = () => {
+export const Notices = () => {
 	const { notices, setNotice } = useOdieAssistantContext();
 
 	if ( Object.keys( notices ).length === 0 ) {
@@ -28,7 +28,7 @@ export const OdieNotices = () => {
 	return (
 		<div className="odie-notices">
 			{ Object.entries( notices ).map( ( [ noticeId, noticeContent ] ) => (
-				<OdieNotice
+				<Notice
 					key={ noticeId }
 					content={ noticeContent }
 					onClose={ () => setNotice( noticeId, null ) }
