+++ packages/odie-client/src/components/notices/style.scss
@@ -2,10 +2,9 @@
 	display: flex;
 	flex-direction: column;
 	gap: 8px;
-	padding: 0 16px;
 	font-size: 0.75rem;
 	position: absolute;
-	bottom: 86px;
+	bottom: 73px;
 
 	.odie-notice {
 		display: flex;
